Handle locale in .uischema. Add check for missing $schema.

+++ packages/dialog/src/library/schemaMerger.ts
@@ -80,16 +80,6 @@ function mergeObjects(obj1: any, obj2: any): any {
     return finalTarget
 }
 
-function fileBase(filename: string): string {
-    return filename.substring(0, filename.indexOf('.'))
-}
-
-function fileLocale(filename: string): string {
-    let extPos = filename.indexOf('.')
-    let localePos = filename.indexOf('.', extPos + 1)
-    return localePos > 0 ? filename.substring(extPos + 1, localePos) : ''
-}
-
 // Build a tree of component (project or package) references in order to compute a topological sort.
 class Component {
     // Name of component
@@ -216,13 +206,12 @@ export default class SchemaMerger {
     private readonly warn: any
     private readonly error: any
     private readonly extensions: string[]
+    private readonly schemaPath: string | undefined
     private readonly debug: boolean | undefined
+    private nugetRoot = ''    // Root where nuget packages are found
 
     // Track packages that have been processed
     private readonly packages = new Set()
-
-    // Root where nuget packages are found
-    private nugetRoot = ''
 
     // Component tree
     private readonly root = new Component()
@@ -273,10 +262,11 @@ export default class SchemaMerger {
      * @param warn Logger for warning messages.
      * @param error Logger for error messages.
      * @param extensions Extensions to analyze for loader.
+     * @param schema Path to merged .schema is doin
      * @param debug Generate debug output.
      * @param nugetRoot Root directory for nuget packages.  (Useful for testing.)
      */
-    public constructor(patterns: string[], output: string, verbose: boolean, log: any, warn: any, error: any, extensions?: string[], debug?: boolean, nugetRoot?: string) {
+    public constructor(patterns: string[], output: string, verbose: boolean, log: any, warn: any, error: any, extensions?: string[], schema?: string, debug?: boolean, nugetRoot?: string) {
         this.patterns = patterns
         this.output = output ? ppath.join(ppath.dirname(output), ppath.basename(output, '.schema')) : ''
         this.verbose = verbose
@@ -284,6 +274,7 @@ export default class SchemaMerger {
         this.warn = warn
         this.error = error
         this.extensions = extensions || ['.schema', '.lu', '.lg', '.qna', '.dialog', '.uischema']
+        this.schemaPath = schema
         this.debug = debug
         this.nugetRoot = nugetRoot || ''
     }
@@ -343,6 +334,18 @@ export default class SchemaMerger {
     private async mergeSchemas(): Promise<any> {
         let fullSchema: any
 
+        if (this.schemaPath) {
+            // Passed in merged schema
+            this.currentFile = this.schemaPath
+            let schema = await fs.readJSON(this.schemaPath)
+            this.currentFile = schema.$schema
+            this.metaSchema = await getJSON(schema.$schema)
+            this.validator.addSchema(this.metaSchema, 'componentSchema')
+            this.currentFile = this.schemaPath
+            this.validateSchema(schema)
+            return parser.dereference(schema)
+        }
+
         // Delete existing output
         await fs.remove(this.output + '.schema')
         await fs.remove(this.output + '.schema.final')
@@ -377,8 +380,10 @@ export default class SchemaMerger {
                         delete component.$ref
                     }
 
-                    // Pick up meta-schema from first .dialog file
-                    if (!this.metaSchema) {
+                    if (!component.$schema) {
+                        this.missingSchemaError()
+                    } else if (!this.metaSchema) {
+                        // Pick up meta-schema from first .dialog file
                         this.metaSchemaId = component.$schema
                         this.currentFile = this.metaSchemaId
                         this.metaSchema = await getJSON(component.$schema)
@@ -485,13 +490,20 @@ export default class SchemaMerger {
         let uiSchemas = this.files.get('.uischema')
         let result = {}
         if (uiSchemas) {
+            if (!schema || this.failed) {
+                this.error('Error must have a merged .schema to merge .uischema files')
+                return
+            }
+
             this.log('Merging component .uischema files')
+            if (this.schemaPath) {
+                this.log(`Using merged schema ${this.schemaPath}`)
+            }
             let outputName = ppath.basename(this.output)
             for (let [fileName, componentPaths] of uiSchemas.entries()) {
                 // Skip files that match output .uischema
                 if (!fileName.startsWith(outputName + '.')) {
-                    let kindName = fileBase(fileName)
-                    let localeName = fileLocale(fileName)
+                    let [kindName, localeName] = this.kindAndLocale(fileName, schema)
                     let locale = result[localeName]
                     if (!locale) {
                         locale = result[localeName] = {}
@@ -511,9 +523,10 @@ export default class SchemaMerger {
                                 this.vlog(`Parsing ${this.currentFile}`)
                             }
                             let component = await fs.readJSON(path)
-
-                            // Pick up meta-schema from first .uischema file
-                            if (!this.metaUISchema) {
+                            if (!component.$schema) {
+                                this.missingSchemaError()
+                            } else if (!this.metaUISchema) {
+                                // Pick up meta-schema from first .uischema file
                                 this.metaUISchemaId = component.$schema
                                 this.currentFile = this.metaUISchemaId
                                 this.metaUISchema = await getJSON(component.$schema)
@@ -551,9 +564,22 @@ export default class SchemaMerger {
         }
     }
 
+    private kindAndLocale(filename: string, schema: any): [string, string] {
+        let kindName = ppath.basename(filename, '.uischema')
+        let locale = ''
+        if (!schema.definitions[kindName]) {
+            let split = kindName.lastIndexOf('.')
+            if (split >= 0) {
+                locale = kindName.substring(split + 1)
+                kindName = kindName.substring(0, split)
+            }
+        }
+        return [kindName, locale]
+    }
+
     // For C# copy all assets into generated/<package>/
     private async copyAssets(): Promise<void> {
-        if (!this.failed) {
+        if (!this.failed && !this.schemaPath) {
             let isCS = false
             for (let component of this.components) {
                 if (component.path.endsWith('.csproj') || component.path.endsWith('.nuspec')) {
@@ -1447,6 +1473,12 @@ export default class SchemaMerger {
             this.missingKinds.add(kind)
             this.failed = true
         }
+    }
+
+    // Missing $schema
+    private missingSchemaError() {
+        this.error(`${this.currentFile}: Error missing $schema`)
+        this.failed = true
     }
 
     // Error in schema validity
